refactor(order): trasform shared commands into requests
