--- order-shared/src/commonMain/kotlin/com/munchies/order/application/port/inbound/command/AdvanceOrderStatusCommand.kt
@@ -1,3 +1,0 @@
-package com.munchies.order.application.port.inbound.command
-
-data class AdvanceOrderStatusCommand(val orderId: String)
