--- order-shared/src/commonMain/kotlin/com/munchies/order/application/port/inbound/command/DiscardOrderCommand.kt
@@ -1,6 +1,0 @@
-package com.munchies.order.application.port.inbound.command
-
-data class DiscardOrderCommand(
-  val orderId: String,
-  val customerId: String,
-)
