--- order-shared/src/commonMain/kotlin/com/munchies/order/application/port/inbound/command/UpdateTakeawayOrderCommand.kt
@@ -1,8 +1,0 @@
-package com.munchies.order.application.port.inbound.command
-
-data class UpdateTakeawayOrderCommand(
-  val orderId: String,
-  val customerId: String,
-  val pickupTime: Long,
-  val customerName: String,
-)
