+++ order-shared/src/commonMain/kotlin/com/munchies/order/infrastructure/adapter/inbound/request/AdvanceOrderStatusRequest.kt
@@ -1,0 +1,3 @@
+package com.munchies.order.infrastructure.adapter.inbound.request
+
+data class AdvanceOrderStatusRequest(val orderId: String)
