+++ order-shared/src/commonMain/kotlin/com/munchies/order/infrastructure/adapter/inbound/request/DiscardOrderRequest.kt
@@ -1,0 +1,6 @@
+package com.munchies.order.infrastructure.adapter.inbound.request
+
+data class DiscardOrderRequest(
+  val orderId: String,
+  val customerId: String,
+)
