+++ order-shared/src/commonMain/kotlin/com/munchies/order/infrastructure/adapter/inbound/request/GetOrderDetailsRequest.kt
@@ -1,0 +1,3 @@
+package com.munchies.order.infrastructure.adapter.inbound.request
+
+data class GetOrderDetailsRequest(val orderId: String)
