+++ order-shared/src/commonMain/kotlin/com/munchies/order/infrastructure/adapter/inbound/request/UpdateOrderItemsRequest.kt
@@ -1,8 +1,8 @@
-package com.munchies.order.application.port.inbound.command
+package com.munchies.order.infrastructure.adapter.inbound.request
 
 import com.munchies.order.infrastructure.adapter.dto.OrderItemDto
 
-data class UpdateOrderItemsCommand(
+data class UpdateOrderItemsRequest(
   val orderId: String,
   val customerId: String,
   val items: List<OrderItemDto>,
