+++ order-shared/src/commonMain/kotlin/com/munchies/order/infrastructure/adapter/inbound/request/UpdateTakeawayOrderRequest.kt
@@ -1,0 +1,8 @@
+package com.munchies.order.infrastructure.adapter.inbound.request
+
+data class UpdateTakeawayOrderRequest(
+  val orderId: String,
+  val customerId: String,
+  val pickupTime: Long,
+  val customerName: String,
+)
